+++ lib/savon_helper.rb
@@ -17,6 +17,10 @@ unless String.respond_to? :blank
     def blank?
       empty?()
     end
+
+    def strip_comments
+      self.match(/^([^#]*)(#.*$)?/).to_a[1]
+    end
   end
 
   class NilClass
@@ -33,4 +37,15 @@ class Object
     name.split('::').last || ''
   end
 
+end
+
+module Kernel
+  alias :oldwarn :warn
+
+  def warn (msg = "", fulltrace = false)
+    trace = caller(1)
+    where = trace[0].sub(/:in.*/, '')
+    $stderr.puts "#{where}: Warning: #{msg}"
+    $stderr.puts trace.map { |t| "\tfrom #{t}" } if fulltrace
+  end
 end