+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## CHANGELOG (notable new features or fixes)
 
+### 0.0.21
+
+* Extended `-fields` to accept a filename to read the fields from
+* Added warning for unparseable entries
+
+
 ### 0.0.20
 
 * `dsc` command refactoring
@@ -9,6 +15,7 @@
 * Extracted savon TypeMapping functionality/DSL into seperate classes
 * Added `--detail_level` flag for `host_detail` command
 * Added `-time_format` flag to specifiy time format
+
 
 ### 0.0.19